[#12] feat: CardType 별도의 파일로 분리
